0.0.0.10
Changed the layout of the Testing Cell as well as actually calling it Testing Cell in engine. Also added a script that adds a smoothed camera follow so it keeps the player in frame at all times. Also zoomed out the camera to better reflect the FoV required for the gameplay this will have.