0.0.0.14
Added the player weapon into the game and made it rotate to follow the cursor. Still need to make it flip on the Y axis when to the left of the Player, as well as make the player flip on the X axis when the cursor is to the left of the Player.